requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests
pandas
llama-index==0.10.0
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
llama-index-vector-stores-qdrant==0.2.8
llama-index-llms-openai-like==0.1.3
openai==1.30.2
pypdf==3.17.2
transformers==4.41.1
gunicorn==22.0.0
llama-index-retrievers-bm25==0.1.3
llama-index-embeddings-huggingface==0.2.0
sentence-transformers==2.7.0
llama-index-readers-file==0.1.1
lxml==5.2.2
nest-asyncio
python-dotenv==1.0.0
